chore(eval): abstract out shared components in eval pipeline
